refactor: custom decorator 타입 올바르게 지정 및 폴더구조 수정
