+++ src/decorators/index.ts
@@ -1,72 +1,1 @@
-import type { Request } from 'express';
-
-export function Middleware(
-  target: any,
-  propertyKey: string,
-  descriptor: PropertyDescriptor,
-) {
-  const method = descriptor.value;
-
-  descriptor.value = function (req: Request) {
-    const args = target.extractors.map(({ type, value }) => {
-      switch (type) {
-        case 'query':
-          return req.query[value];
-        case 'request':
-          return req;
-        case 'cookies':
-          return req.cookies[value];
-        case 'authorization':
-          if (!req.headers.authorization) {
-            return null;
-          }
-
-          return req.headers.authorization.split(`${value} `)[1];
-      }
-    });
-
-    return method(...args);
-  };
-}
-
-export function Query(property: string) {
-  // eslint-disable-next-line
-  return function (target: any, propertyKey: string, parameterIndex: number) {
-    if (!target.extractors) {
-      target.extractors = [];
-    }
-
-    target.extractors.unshift({ type: 'query', value: property });
-  };
-}
-
-// eslint-disable-next-line
-export function Req(target: any, propertyKey: string, parameterIndex: number) {
-  if (!target.extractors) {
-    target.extractors = [];
-  }
-
-  target.extractors.unshift({ type: 'request', value: null });
-}
-
-export function Cookies(key: string) {
-  // eslint-disable-next-line
-  return function (target: any, propertyKey: string, parameterIndex: number) {
-    if (!target.extractors) {
-      target.extractors = [];
-    }
-
-    target.extractors.unshift({ type: 'cookies', value: key });
-  };
-}
-
-export function Auth(type: string) {
-  // eslint-disable-next-line
-  return function (target: any, propertyKey: string, parameterIndex: number) {
-    if (!target.extractors) {
-      target.extractors = [];
-    }
-
-    target.extractors.unshift({ type: 'authorization', value: type });
-  };
-}
+export * from './middlewear.decorator';
